Public API

MillerIndices.Miller — Type

Miller(i, j, k)

Represent the Miller indices in the real space (crystal directions).

MillerIndices.MillerBravais — Type

MillerBravais(i, j, k, l)

Represent the Miller–Bravais indices in the real space (crystal directions).

MillerIndices.ReciprocalMiller — Type

ReciprocalMiller(i, j, k)

Represent the Miller indices in the reciprocal space (planes).

MillerIndices.ReciprocalMillerBravais — Type

ReciprocalMillerBravais(i, j, k, l)

Represent the Miller–Bravais indices in the reciprocal space (planes).

MillerIndices.listfamily — Function

listfamily(x::Union{Miller,MillerBravais,ReciprocalMiller,ReciprocalMillerBravais})

List the all the directions/planes that are equivalent to x by symmetry.

MillerIndices.@m_str — Macro

m_str(s)

Generate the Miller indices or Miller–Bravais indices quickly.

Examples

julia> m"[-1, 0, 1]"
3-element Miller:
 -1
  0
  1

julia> m"<2, -1, -1, 3>"
4-element MillerBravais:
  2
 -1
 -1
  3

julia> m"(-1, 0, 1)"
3-element ReciprocalMiller:
 -1
  0
  1

julia> m"(1, 0, -1, 0)"
4-element ReciprocalMillerBravais:
  1
  0
 -1
  0